fix: Delete the custom file and import the node_modules file instead Modify the file directory structure add Page refreshes when switching languages
