+++ frontend/public/index.html
@@ -26,7 +26,7 @@
 </head>
 
 <body>
-  <link rel="stylesheet/less" type="text/css" href="/theme.less" />
+  <link rel="stylesheet/less" type="text/css" href="%PUBLIC_URL%/antd/theme.less" />
   <noscript>You need to enable JavaScript to run this app.</noscript>
   <div id="root"></div>
   <script>
@@ -36,7 +36,7 @@
     };
 
   </script>
-  <script type="text/javascript" src="/less.min.js"></script>
+  <script type="text/javascript" src="%PUBLIC_URL%/antd/less.min.js"></script>
   <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
