--- frontend/src/styles/theme/light.js
@@ -1,69 +1,0 @@
-module.exports = {
-  'alert-error-bg-color': "~`colorPalette('@{error-color}', 1) `",
-  'alert-error-border-color': "~`colorPalette('@{error-color}', 3) `",
-  'alert-info-bg-color': "~`colorPalette('@{info-color}', 1) `",
-  'alert-info-border-color': "~`colorPalette('@{info-color}', 3) `",
-  'alert-success-bg-color': "~`colorPalette('@{success-color}', 1) `",
-  'alert-success-border-color': "~`colorPalette('@{success-color}', 3) `",
-  'alert-warning-bg-color': "~`colorPalette('@{warning-color}', 1) `",
-  'alert-warning-border-color': "~`colorPalette('@{warning-color}', 3) `",
-  'avatar-bg': '#e9ecef',
-  'avatar-color': '#6c757d',
-  'background-color-base': 'hsv(0, 0, 96%) ',
-  'background-color-light': 'hsv(0, 0, 98%) ',
-  'badge-text-color': '@component-background',
-  'body-background': '#fff',
-  'border-color-base': 'hsv(0, 0, 85%) ',
-  'border-color-split': 'hsv(0, 0, 94%) ',
-  'border-style-base': 'solid ',
-  'btn-default-bg': '@component-background',
-  'card-actions-background': '@background-color-light',
-  'card-shadow':
-    '0 1px 2px -2px rgba(0, 0, 0, 0.16), 0 3px 6px 0 rgba(0, 0, 0, 0.12),',
-  'component-background': '#fff',
-  'disabled-color': 'fade(#000, 25%)',
-  'disabled-color-dark': 'fade(#fff, 35%)',
-  'font-family':
-    "-apple-system, BlinkMacSystemFont, 'Segoe UI', 'PingFang SC', 'Hiragino Sans GB',",
-  'font-feature-settings-base': "'tnum'",
-  'heading-color': 'fade(#000, 85%)',
-  'icon-color-hover': 'fade(@black, 75%)',
-  'input-bg': '@component-background',
-  'input-number-handler-active-bg': '#f4f4f4',
-  'input-placeholder-color': 'hsv(0, 0, 75%)',
-  'item-hover-bg': '#f5f5f5',
-  'layout-body-background': '#f0f2f5',
-  'layout-header-background': '#001529',
-  'layout-sider-background': '@layout-header-background',
-  'layout-trigger-background': '#002140',
-  'menu-dark-submenu-bg': '#000c17',
-  'menu-popup-bg': '@component-background',
-  'modal-header-bg': '@component-background',
-  'outline-blur-size': '0px',
-  'pagination-item-bg-active': '@component-background',
-  'popover-bg': '@component-background',
-  'select-item-selected-font-weight': '600px',
-  'shadow-1-down':
-    '0 6px 16px -8px rgba(0, 0, 0, 0.08), 0 9px 28px 0 rgba(0, 0, 0, 0.05),',
-  'shadow-1-left':
-    '-6px 0 16px -8px rgba(0, 0, 0, 0.08), -9px 0 28px 0 rgba(0, 0, 0, 0.05),',
-  'shadow-1-right':
-    '6px 0 16px -8px rgba(0, 0, 0, 0.08), 9px 0 28px 0 rgba(0, 0, 0, 0.05),',
-  'shadow-1-up':
-    '0 -6px 16px -8px rgba(0, 0, 0, 0.08), 0 -9px 28px 0 rgba(0, 0, 0, 0.05),',
-  'shadow-2':
-    '0 3px 6px -4px rgba(0, 0, 0, 0.12), 0 6px 16px 0 rgba(0, 0, 0, 0.08),',
-  'shadow-color': 'rgba(0, 0, 0, 0.15)',
-  'skeleton-color': '#f2f2f2',
-  'slider-handle-color': '@primary-3',
-  'slider-rail-background-color': '@background-color-base',
-  'slider-rail-background-color-hover': '#e1e1e1',
-  'slider-track-background-color': '@primary-3',
-  'table-body-sort-bg': 'rgba(0, 0, 0, 0.01)',
-  'table-header-bg': '@background-color-light',
-  'table-header-sort-bg': '@background-color-base',
-  'table-row-hover-bg': '@item-hover-bg',
-  'text-color': 'fade(@black, 65%)',
-  'text-color-secondary': 'fade(@black, 45%)',
-  'tooltip-bg': 'rgba(0, 0, 0, 0.75)',
-};
